requirements.txt
requests==2.31.0
pandas==2.1.1
openpyxl==3.1.2
xlsxwriter==3.1.9
docopt==0.6.2
Jinja2==3.1.2
python-dotenv==1.0.1